+++ internal/provider/provider.go
@@ -52,6 +52,7 @@ func (p *utilsProvider) Functions(ctx context.Context) []func() function.Functio
 		NewNormalizeMaskFunction,
 		NewNormalizeMacFunction,
 		NewMergeFunction,
+		NewParseBgpRdRtFunction,
 	}
 }